+++ crates/libsy/src/algorithms/escalation.rs
@@ -1,29 +1,27 @@
 // SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-//! Trajectory-judged escalation: serve the efficient model until an LLM judge finds the run
-//! in trouble, then latch the session to the capable model for the rest of the task.
+//! Trajectory-judged escalation: serve the efficient model until an LLM judge finds the run in
+//! trouble, then latch the session to the capable model for the rest of the task — re-trying a
+//! model that has already failed the task wastes a turn.
 //!
-//! Each turn is served on the efficient tier first. The judge then reads the conversation
-//! *including that reply* and decides whether the run is in trouble. A run that is fine keeps
-//! the reply it already has, so the common turn costs one serving call plus the judge; only a
-//! confirmed escalation pays for a second serving call, on the capable model.
+//! Each turn is served on the efficient tier first, and the judge reads the conversation
+//! *including that reply*. A run that is fine keeps the reply it already has, so only a confirmed
+//! escalation pays for a second serving call. A judge failure fails open to the efficient tier
+//! and never latches: an outage costs quality risk, never money.
 //!
-//! Once a session escalates it stays on the capable model, since re-trying a model that has
-//! already failed this task wastes a turn. A judge failure fails open to the efficient tier and
-//! never latches: an outage costs quality risk, never money.
-//!
-//! The algorithm is a [`FallThrough`] composition: the session latch and the escalation judge
-//! are its classifiers, and the confirmation streak is folded in by a processor once the turn's
-//! decision is final.
+//! Assembled as a [`FallThrough`]: the latch and the judge are its classifiers, and a processor
+//! commits the confirmation streak once the turn's decision is final.
 
 use std::sync::Arc;
 
 use async_trait::async_trait;
 use serde::Deserialize;
-use switchyard_protocol::{
-    AggLlmResponse, ContentBlock, LlmRequest, LlmResponse, Message, OutputParams, Role,
-};
+use switchyard_protocol::{AggLlmResponse, LlmRequest, LlmResponse, Message, OutputParams, Role};
+
+mod transcript;
+
+use transcript::{conversation_turn, summarize_for_judge};
 
 use crate::{
     algorithms::util::{
@@ -49,53 +47,28 @@ const STREAK_KEY: &str = "escalation_streak";
 /// travels from the classifier to [`ConfirmationProcessor`].
 const PENDING_KEY: &str = "escalation_pending_streak";
 
-/// Separator marking where [`truncate_middle`] dropped a message's interior.
-const TRIM_MARKER: &str = " ...[trimmed] ";
-
-/// Suffix marking a transcript cut off by [`MAX_REQUEST_CHARS`].
-const TRUNCATION_SUFFIX: &str = "...<truncated>";
-
-/// Per-message cap for system and developer anchors. Coding-agent harnesses inject very large
-/// boilerplate system prompts carrying no trajectory signal; uncapped they crowd out the window.
-const SYSTEM_CHARS: usize = 1_000;
-
-/// Cap for the first user message — the task statement the judge needs to detect drift, so it
-/// gets the most generous anchor budget.
-const FIRST_USER_CHARS: usize = 2_000;
-
-/// Backstop on the whole assembled transcript, for a pathological single message. The window
-/// caps below normally bind first; when this does bind, the oldest window lines drop.
-const MAX_REQUEST_CHARS: usize = 18_000;
-
-/// Completion budget for one judge reply, covering any reasoning the judge model emits
-/// alongside the verdict.
+/// Completion budget for one judge reply, covering any reasoning emitted alongside the verdict.
 ///
-/// A runaway guard, not a budget: output tokens cost what they generate, so a generous cap is
-/// nearly free, while a tight one truncates mid-reasoning into unparseable JSON and fails the
-/// judge open on every call. Sized well above the verdict itself — the benchmarked judge's
-/// `reason` ran ~40 tokens at the median and ~136 at its longest — leaving the remainder as
-/// reasoning headroom.
+/// A runaway guard, not a budget: a tight cap truncates mid-reasoning into unparseable JSON and
+/// fails the judge open on every call, while a generous one costs only what it generates. The
+/// benchmarked judge's `reason` ran ~40 tokens at the median and ~136 at its longest.
 const JUDGE_MAX_OUTPUT_TOKENS: u64 = 4_096;
 
 /// The tuning surface for the trajectory judge.
 ///
-/// Deliberately small: these are the three settings an audit of the Python router's twenty-odd
-/// knobs found to actually change routing outcomes. Everything else it exposed is either a
-/// fixed invariant (see the `*_CHARS` constants above) or operational plumbing.
-///
-/// Defaults are the benchmarked configuration; omitted fields take those values.
+/// Deliberately small: an audit of the Python router's twenty-odd knobs found only these three
+/// change routing outcomes; the rest were fixed invariants (the `*_CHARS` constants above) or
+/// operational plumbing. Defaults are the benchmarked configuration.
 #[derive(Clone, Debug, Deserialize)]
 #[serde(default)]
 pub struct EscalationJudgeSettings {
     /// Consecutive escalate verdicts required before the latch fires.
     ///
-    /// `1` latches on the first verdict. Higher values filter one-shot eager verdicts — a
-    /// single failed command misread as a pattern — while keeping recall on real trouble,
-    /// which by definition persists across turns. On the benchmarked workload `2` suppressed
-    /// roughly two thirds of escalate verdicts, so this is the router's main cost dial.
-    ///
-    /// The streak is keyed on the session id and held per process. A request without a
-    /// session id cannot accumulate one, so at `2` or higher it can never escalate.
+    /// `1` latches on the first verdict; higher values filter one-shot eager verdicts — a single
+    /// failed command misread as a pattern — while keeping recall on trouble that persists. On
+    /// the benchmarked workload `2` suppressed roughly two thirds of escalate verdicts, making
+    /// this the main cost dial. The streak is keyed on the session id, so a request without one
+    /// can never escalate at `2` or higher.
     pub confirmations: u32,
     /// Trailing messages shown on top of the anchors. Loop detection needs to see the
     /// repeats, so a cycle longer than this window is invisible to the judge.
@@ -199,11 +172,9 @@ impl Judge for EscalationJudge {
     }
 }
 
-/// Maps the judge's verdict to a routing classification, keeping "the judge declined" and
-/// "the judge was unavailable" distinguishable.
-///
-/// [`Classification::Ambiguous`] carries the unavailable case: it argmaxes to nothing, exactly
-/// like an empty score set, but the classifier can tell them apart when recording the verdict.
+/// Carries the judge's three outcomes as a [`Classification`]. `Ambiguous` marks "judge
+/// unavailable": it argmaxes to nothing exactly like a decline's empty score set, but the
+/// classifier can still tell the two apart.
 struct EscalationPolicy {
     capable: String,
 }
@@ -225,9 +196,8 @@ impl JudgePolicy for EscalationPolicy {
 
 /// The session latch, labelled with escalation's tiers.
 ///
-/// [`AffinityRouter`] is model-agnostic and cannot know which of its targets is the capable
-/// one, so a latched turn would otherwise be routed with no tier label — and after escalation
-/// that is every remaining turn of the session.
+/// [`AffinityRouter`] cannot know which of its targets is the capable one, so a latched turn
+/// would otherwise route untiered — and after escalation that is every remaining turn.
 struct LatchedTier {
     affinity: AffinityRouter,
     capable: String,
@@ -286,16 +256,15 @@ impl Classifier<State> for EscalationClassifier {
         request: &mut Request,
         driver: Option<&Driver>,
     ) -> Result<Scored> {
-        // A missing driver is a broken composition: this classifier cannot decide without
-        // calling a model.
+        // A broken composition, not a routing condition: this cannot decide without a model.
         let Some(driver) = driver else {
             return Err(LibsyError::AlgorithmError {
                 message: "escalation classifier requires a driver to call its targets".to_string(),
             });
         };
 
-        // Serve the turn on the efficient tier. The judge reads this reply, so the call comes
-        // before the verdict — and on the common path the reply is also the turn's answer.
+        // The judge reads this reply, so the call comes before the verdict — and on the common
+        // path the reply is also the turn's answer.
         let served = driver
             .call_llm_target(
                 Context::default(),
@@ -306,8 +275,7 @@ impl Classifier<State> for EscalationClassifier {
                 }),
             )
             .await?;
-        // Judging the reply means buffering it: a stream cannot be read twice, and the judge
-        // has to read it before the caller can be given it.
+        // Judging the reply means buffering it: a stream cannot be read twice.
         let Response {
             llm_response,
             metadata,
@@ -316,8 +284,7 @@ impl Classifier<State> for EscalationClassifier {
             LibsyError::external("aggregating the efficient tier response", error)
         })?;
 
-        // Judge the trajectory with this turn's reply appended, so the newest evidence is the
-        // work actually being judged.
+        // Append this turn's reply, so the newest evidence is the work being judged.
         let mut judged = request.clone();
         judged
             .llm_request
@@ -325,10 +292,9 @@ impl Classifier<State> for EscalationClassifier {
             .push(assistant_message(&aggregate));
         let (classification, _) = self.judge.score(state, &mut judged, Some(driver)).await?;
 
-        // Fold the verdict into the streak this turn would leave behind. Strict-consecutive:
-        // an escalate verdict extends the streak, any decline clears it, and an unavailable
-        // judge is no evidence either way so it leaves the streak untouched. The processor
-        // commits `pending` once the turn's decision is final.
+        // The streak this turn would leave behind, committed by the processor once the decision
+        // is final. Strict-consecutive: escalate extends it, any decline clears it, and an
+        // unavailable judge is no evidence either way.
         let held = streak(state);
         let (escalate, pending) = match classification {
             Classification::Scores(ref scores) if !scores.is_empty() => (true, held + 1),
@@ -355,8 +321,8 @@ impl Classifier<State> for EscalationClassifier {
 
 /// Commits the streak the classifier computed for this turn.
 ///
-/// Split from the classifier so the streak only moves once the turn's decision is final: the
-/// classifier reads the streak to gate its verdict, this writes it back afterwards.
+/// Split from the classifier so the streak only moves once the decision is final: the classifier
+/// reads it to gate its verdict, this writes it back afterwards.
 struct ConfirmationProcessor;
 
 #[async_trait]
@@ -407,10 +373,8 @@ fn assistant_message(response: &AggLlmResponse) -> Message {
     }
 }
 
-/// The escalation router's decision reasoning.
-///
-/// The cascade reports only the winning score, so this names the tier rather than the path
-/// that chose it; whether a turn was latched or judged is visible in the call's tier field.
+/// Names the tier rather than the path that chose it: the cascade reports only the winning
+/// score, and latched-vs-judged is already visible in the call's tier field.
 fn decision_reason(_name: &str, winner: &Score) -> String {
     format!("escalation selected {}", winner.target)
 }
@@ -471,8 +435,8 @@ impl EscalationRouter {
 
         // The capable target leads the set so `count_tokens_client` prefers its client.
         let targets = LlmTargetSet::new(vec![capable_target, efficient_target]);
-        // The latch classifies first: a session that already escalated skips the judge, and
-        // the efficient call the judge would need, entirely.
+        // The latch classifies first, so an escalated session skips both the judge and the
+        // efficient call it would need.
         let route = FallThrough::<State>::new_with_state(targets)
             .with_name(ALGORITHM_NAME)
             .with_decision_reason(decision_reason)
@@ -503,149 +467,6 @@ impl Algorithm for EscalationRouter {
     }
 }
 
-/// The 1-indexed model invocation the transcript ends on: one per assistant reply, the
-/// newest being the turn under judgement.
-///
-/// libsy receives messages already normalized by `switchyard-protocol`, so unlike the
-/// wire-format-aware Python equivalent this needs no per-format branching.
-fn conversation_turn(request: &Request) -> usize {
-    request
-        .llm_request
-        .messages
-        .iter()
-        .filter(|message| message.role == Role::Assistant)
-        .count()
-}
-
-/// Flattens a message to plain text, tool calls and tool results included.
-///
-/// [`Message::text_content`] is deliberately not used here: it keeps only text and refusal
-/// blocks, which would erase exactly the repeated-command signal the judge's loop detection
-/// relies on.
-fn message_text(message: &Message) -> String {
-    let mut parts = Vec::new();
-    collect_text(&message.content, &mut parts);
-    parts.join(" ")
-}
-
-/// Appends the judge-relevant text of each block, descending into tool results.
-fn collect_text(content: &[ContentBlock], parts: &mut Vec<String>) {
-    for block in content {
-        match block {
-            ContentBlock::Text { text } | ContentBlock::Refusal { text } => {
-                parts.push(text.clone());
-            }
-            ContentBlock::ToolCall(call) => {
-                parts.push(format!("tool_call {}({})", call.name, call.arguments));
-            }
-            ContentBlock::ToolResult(result) => collect_text(&result.content, parts),
-            _ => {}
-        }
-    }
-}
-
-/// Keeps the head and tail of `text` within `limit` characters.
-///
-/// The head gets two thirds of the surviving budget: for a trajectory judge the command or
-/// error signature that opens a message carries more signal than its trailing output.
-fn truncate_middle(text: &str, limit: usize) -> String {
-    let chars: Vec<char> = text.chars().collect();
-    if chars.len() <= limit {
-        return text.to_string();
-    }
-    let keep = limit
-        .saturating_sub(TRIM_MARKER.chars().count())
-        .max(20)
-        .min(chars.len());
-    let head = keep * 2 / 3;
-    let tail = keep - head;
-    let mut out: String = chars[..head].iter().collect();
-    out.push_str(TRIM_MARKER);
-    out.extend(chars[chars.len() - tail..].iter());
-    out
-}
-
-/// Renders a compact role-labelled transcript for the judge.
-///
-/// The framing anchors — system/developer messages and the first user message, where agent
-/// harnesses put the task statement — are kept unconditionally and capped individually. The
-/// trailing window carries recent activity. A coverage header states how much history is not
-/// shown, so the judge can reason about pace rather than assuming it sees everything.
-///
-/// When the assembled text still exceeds `max_request_chars`, the oldest window lines go
-/// first: for a trajectory judge the newest evidence is strictly the most valuable.
-fn summarize_for_judge(
-    messages: &[Message],
-    turn: usize,
-    settings: &EscalationJudgeSettings,
-) -> String {
-    let mut anchors: Vec<String> = Vec::new();
-    let mut window: Vec<String> = Vec::new();
-    let mut first_user_seen = false;
-
-    for message in messages {
-        let text = message_text(message);
-        match message.role {
-            Role::System | Role::Developer => anchors.push(format!(
-                "[{}] {}",
-                role_label(message.role),
-                truncate_middle(&text, SYSTEM_CHARS)
-            )),
-            Role::User if !first_user_seen => {
-                first_user_seen = true;
-                anchors.push(format!(
-                    "[user (task)] {}",
-                    truncate_middle(&text, FIRST_USER_CHARS)
-                ));
-            }
-            role => window.push(format!(
-                "[{}] {}",
-                role_label(role),
-                truncate_middle(&text, settings.window_message_chars)
-            )),
-        }
-    }
-
-    if window.len() > settings.recent_turn_window {
-        window.drain(..window.len() - settings.recent_turn_window);
-    }
-
-    let header = format!(
-        "Conversation turn {turn}; showing the last {} of {} messages after the task framing.",
-        window.len(),
-        messages.len(),
-    );
-    let assemble = |window: &[String]| {
-        std::iter::once(header.as_str())
-            .chain(anchors.iter().map(String::as_str))
-            .chain(window.iter().map(String::as_str))
-            .collect::<Vec<_>>()
-            .join("\n")
-    };
-
-    let mut text = assemble(&window);
-    while text.chars().count() > MAX_REQUEST_CHARS && !window.is_empty() {
-        window.remove(0);
-        text = assemble(&window);
-    }
-    if text.chars().count() > MAX_REQUEST_CHARS {
-        let keep = MAX_REQUEST_CHARS.saturating_sub(TRUNCATION_SUFFIX.chars().count() + 1);
-        text = text.chars().take(keep).collect::<String>() + TRUNCATION_SUFFIX;
-    }
-    text
-}
-
-/// The transcript label for a role.
-fn role_label(role: Role) -> &'static str {
-    match role {
-        Role::System => "system",
-        Role::Developer => "developer",
-        Role::User => "user",
-        Role::Assistant => "assistant",
-        Role::Tool => "tool",
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::collections::VecDeque;
@@ -653,10 +474,9 @@ mod tests {
 
     use futures::StreamExt;
     use parking_lot::Mutex;
-    use serde_json::json;
     use switchyard_protocol::{
-        completion_text, text_response, ContentBlock, LlmClientError, LlmRequest, LlmResponseChunk,
-        Message, Metadata, Role, ToolCall, ToolResult,
+        completion_text, text_response, LlmClientError, LlmRequest, LlmResponseChunk, Message,
+        Metadata, Role,
     };
 
     use crate::{
@@ -664,9 +484,8 @@ mod tests {
     };
 
     use super::{
-        conversation_turn, load_judge_config, message_text, summarize_for_judge, truncate_middle,
-        EscalationJudge, EscalationJudgeSettings, EscalationRouter, Judge, State,
-        JUDGE_MAX_OUTPUT_TOKENS, MAX_REQUEST_CHARS, PROMPT_TEMPLATE, SCHEMA_TEMPLATE,
+        load_judge_config, EscalationJudge, EscalationJudgeSettings, EscalationRouter, Judge,
+        State, JUDGE_MAX_OUTPUT_TOKENS, PROMPT_TEMPLATE, SCHEMA_TEMPLATE,
     };
 
     fn target(name: &str, client: Option<Arc<dyn RoutedLlmClient>>) -> LlmTarget {
@@ -817,6 +636,23 @@ mod tests {
         Ok((router, calls))
     }
 
+    /// Drives `router` through `turns` requests in one session, returning the last trace.
+    async fn run_turns(
+        router: &Arc<EscalationRouter>,
+        session: Option<&str>,
+        turns: usize,
+    ) -> crate::Result<Vec<Arc<dyn Decision>>> {
+        let mut trace = Vec::new();
+        for _ in 0..turns {
+            trace = router
+                .clone()
+                .run(Context::default(), request(session))
+                .await?
+                .0;
+        }
+        Ok(trace)
+    }
+
     #[tokio::test]
     async fn pass_judge_keeps_the_efficient_reply() -> crate::Result<()> {
         let (router, calls) = instrumented_router([
@@ -868,43 +704,18 @@ mod tests {
             latch_immediately(),
         )?;
 
-        router
-            .clone()
-            .run(Context::default(), request(Some("s3")))
-            .await?;
-        let (trace, _) = router.run(Context::default(), request(Some("s3"))).await?;
+        let trace = run_turns(&router, Some("s3"), 2).await?;
 
         assert_eq!(trace.len(), 1);
         assert_eq!(trace[0].selected_model(), "capable");
+        // The latch is model-agnostic, so escalation labels the tier itself; after a session
+        // escalates this is every remaining turn.
+        assert_eq!(trace[0].routing_tier(), Some("strong"));
         // A latched turn costs one call: no efficient reply for the judge to read, no judge.
         assert_eq!(
             *calls.lock(),
             vec!["efficient", "judge", "capable", "capable"]
         );
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn a_pinned_turn_is_still_labelled_strong() -> crate::Result<()> {
-        let (router, _) = instrumented_router_with(
-            [
-                ok("I give up"),
-                ok(&verdict_json(true)),
-                ok("right"),
-                ok("right again"),
-            ],
-            latch_immediately(),
-        )?;
-
-        router
-            .clone()
-            .run(Context::default(), request(Some("s3b")))
-            .await?;
-        let (trace, _) = router.run(Context::default(), request(Some("s3b"))).await?;
-
-        // The latch is model-agnostic, so escalation labels its tier explicitly; after a
-        // session escalates this is every remaining turn.
-        assert_eq!(trace[0].routing_tier(), Some("strong"));
         Ok(())
     }
 
@@ -976,15 +787,7 @@ mod tests {
             ok("still strong"),      // capable (third request — pinned)
         ])?;
 
-        router
-            .clone()
-            .run(Context::default(), request(Some("c2")))
-            .await?;
-        router
-            .clone()
-            .run(Context::default(), request(Some("c2")))
-            .await?;
-        let (trace, _) = router.run(Context::default(), request(Some("c2"))).await?;
+        let trace = run_turns(&router, Some("c2"), 3).await?;
 
         assert_eq!(trace[0].selected_model(), "capable");
         assert_eq!(
@@ -1012,12 +815,7 @@ mod tests {
             ok(&verdict_json(true)),  // streak 1 again, not 2
         ])?;
 
-        for _ in 0..3 {
-            router
-                .clone()
-                .run(Context::default(), request(Some("c3")))
-                .await?;
-        }
+        run_turns(&router, Some("c3"), 3).await?;
 
         assert!(
             !calls.lock().contains(&"capable".to_string()),
@@ -1038,13 +836,7 @@ mod tests {
             ok("strong answer"),     // capable
         ])?;
 
-        for _ in 0..2 {
-            router
-                .clone()
-                .run(Context::default(), request(Some("c4")))
-                .await?;
-        }
-        let (trace, _) = router.run(Context::default(), request(Some("c4"))).await?;
+        let trace = run_turns(&router, Some("c4"), 3).await?;
 
         assert_eq!(trace[0].selected_model(), "capable");
         Ok(())
@@ -1059,11 +851,7 @@ mod tests {
             ok(&verdict_json(true)),
         ])?;
 
-        router
-            .clone()
-            .run(Context::default(), request(None))
-            .await?;
-        router.run(Context::default(), request(None)).await?;
+        run_turns(&router, None, 2).await?;
 
         // No session id means no retained state, so the capable tier is unreachable at
         // confirmations > 1. Documented behaviour, asserted so it cannot change silently.
@@ -1087,11 +875,7 @@ mod tests {
             latch_immediately(),
         )?;
 
-        router
-            .clone()
-            .run(Context::default(), request(None))
-            .await?;
-        router.run(Context::default(), request(None)).await?;
+        run_turns(&router, None, 2).await?;
 
         assert_eq!(
             *calls.lock(),
@@ -1156,123 +940,5 @@ mod tests {
         );
         assert!(built.llm_request.output.response_format.is_some());
         Ok(())
-    }
-
-    #[test]
-    fn conversation_turn_counts_assistant_replies() {
-        // The transcript handed to the judge ends on the reply being judged, so the count is
-        // the assistant replies present — no lookahead.
-        assert_eq!(conversation_turn(&request_at_turn(None, 1)), 0);
-        assert_eq!(conversation_turn(&request_at_turn(None, 5)), 4);
-    }
-
-    #[test]
-    fn message_text_keeps_tool_calls_and_results() {
-        let call = Message {
-            role: Role::Assistant,
-            content: vec![
-                ContentBlock::Text {
-                    text: "running it".to_string(),
-                },
-                ContentBlock::ToolCall(ToolCall {
-                    id: "call-1".to_string(),
-                    name: "bash".to_string(),
-                    arguments: json!({"cmd": "ls"}),
-                }),
-            ],
-        };
-        let text = message_text(&call);
-        assert!(text.contains("running it"), "{text}");
-        assert!(text.contains(r#"tool_call bash({"cmd":"ls"})"#), "{text}");
-
-        let result = Message {
-            role: Role::Tool,
-            content: vec![ContentBlock::ToolResult(ToolResult {
-                tool_call_id: "call-1".to_string(),
-                content: vec![ContentBlock::Text {
-                    text: "no such file".to_string(),
-                }],
-                is_error: Some(true),
-            })],
-        };
-        assert_eq!(message_text(&result), "no such file");
-    }
-
-    #[test]
-    fn truncate_middle_keeps_head_and_tail() {
-        let text = "a".repeat(40) + &"z".repeat(40);
-        let trimmed = truncate_middle(&text, 50);
-        assert!(trimmed.chars().count() <= 50, "{trimmed}");
-        assert!(trimmed.starts_with('a'));
-        assert!(trimmed.ends_with('z'));
-        assert!(trimmed.contains("[trimmed]"));
-
-        // Under the limit the text is returned untouched.
-        assert_eq!(truncate_middle("short", 50), "short");
-    }
-
-    #[test]
-    fn summary_keeps_anchors_and_the_recent_window() {
-        let mut messages = vec![
-            Message::text(Role::System, "you are a coding agent"),
-            Message::text(Role::User, "fix the failing test"),
-        ];
-        for i in 0..10 {
-            messages.push(Message::text(Role::Assistant, format!("step {i}")));
-        }
-        let settings = EscalationJudgeSettings {
-            recent_turn_window: 3,
-            ..EscalationJudgeSettings::default()
-        };
-
-        let summary = summarize_for_judge(&messages, 10, &settings);
-
-        assert!(
-            summary.contains("[system] you are a coding agent"),
-            "{summary}"
-        );
-        assert!(
-            summary.contains("[user (task)] fix the failing test"),
-            "{summary}"
-        );
-        assert!(summary.contains("Conversation turn 10; showing the last 3 of 12 messages"));
-        // Only the newest window entries survive.
-        assert!(summary.contains("step 9"), "{summary}");
-        assert!(summary.contains("step 7"), "{summary}");
-        assert!(!summary.contains("step 6"), "{summary}");
-    }
-
-    #[test]
-    fn summary_drops_oldest_window_lines_under_the_char_cap() {
-        // MAX_REQUEST_CHARS is a backstop, not a dial: at default settings the window caps
-        // bind first (28 x 500 plus anchors sits under it), so reaching it takes an unusually
-        // wide per-message cap. That is the point — it only fires on pathological input.
-        let mut messages = vec![
-            Message::text(Role::System, "framing"),
-            Message::text(Role::User, "task"),
-        ];
-        for i in 0..20 {
-            messages.push(Message::text(
-                Role::Assistant,
-                format!("{i} {}", "x".repeat(2_000)),
-            ));
-        }
-        let settings = EscalationJudgeSettings {
-            window_message_chars: 2_000,
-            ..EscalationJudgeSettings::default()
-        };
-
-        let summary = summarize_for_judge(&messages, 20, &settings);
-
-        assert!(
-            summary.chars().count() <= MAX_REQUEST_CHARS,
-            "{}",
-            summary.chars().count()
-        );
-        // Anchors are never dropped, and the newest activity outlives the oldest.
-        assert!(summary.contains("[system] framing"), "{summary}");
-        assert!(summary.contains("[user (task)] task"), "{summary}");
-        assert!(summary.contains("19 xxx"), "{summary}");
-        assert!(!summary.contains("0 xxx"), "{summary}");
     }
 }
